sarama message unmarshall test
